# encoding/hex: optimizations to EncodeToString