Currently, a string alloc/copy happens for the result value,
but we can wrap the buffer instead. Additionally, we can skip it all
for empty input.

name                     old time/op   new time/op    delta
EncodeToString/0-10       4.66ns ± 0%    0.31ns ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/256-10      321ns ± 0%     267ns ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/1024-10    1.20µs ± 0%    1.04µs ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/4096-10    4.83µs ± 0%    4.23µs ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/16384-10   18.0µs ± 0%    16.0µs ± 0%   ~     (p=1.000 n=1+1)

name                     old speed     new speed      delta
EncodeToString/256-10    797MB/s ± 0%   960MB/s ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/1024-10   855MB/s ± 0%   983MB/s ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/4096-10   849MB/s ± 0%   969MB/s ± 0%   ~     (p=1.000 n=1+1)
EncodeToString/16384-10  912MB/s ± 0%  1022MB/s ± 0%   ~     (p=1.000 n=1+1)